scp-cleaning-functions/requirements.txt
azure-functions>=1.17
azure-storage-blob>=12.19
pandas>=2.0
numpy>=1.24
openpyxl>=3.1
rapidfuzz>=3.0
python-dateutil>=2.8
ftfy>=6.0
chardet>=5.0
recordlinkage>=0.16
pydantic>=2.0
requests>=2.31
